CDP-1561 - remove components used for hooks refactor testing
