+++ components/ScrollableTableWithMenu/TableActionsMenu/TableActionsMenu.js
@@ -1,9 +1,7 @@
 /* eslint-disable react/destructuring-assignment */
-/* eslint-disable react/prefer-stateless-function */
-import React, { Fragment } from 'react';
+import React, { Fragment, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { graphql } from 'react-apollo';
-import compose from 'lodash.flowright';
+import { useQuery } from '@apollo/react-hooks';
 import { Button, Checkbox, Modal } from 'semantic-ui-react';
 import ApolloError from 'components/errors/ApolloError';
 import editIcon from 'static/images/dashboard/edit.svg';
@@ -17,84 +15,117 @@ import ActionResults from './ActionResults/ActionResults';
 import { TEAM_VIDEO_PROJECTS_QUERY } from '../TableBody/TableBody';
 import { TEAM_VIDEO_PROJECTS_COUNT_QUERY } from '../TablePagination/TablePagination';
 import './TableActionsMenu.scss';
-
 // TEMP
 import { packageMocks, documentFileMocks } from '../TableBody/mocks';
 
-class TableActionsMenu extends React.Component {
-  state = {
-    displayConfirmationMsg: false,
-    deleteConfirmOpen: false,
-    actionFailures: []
-  };
+const CONFIRMATION_MSG_DELAY = 3000;
 
-  _isMounted = false;
+const getDraftProjects = projects => {
+  if ( !projects ) return [];
+  return projects.reduce( ( acc, curr ) => {
+    if ( curr.status === 'DRAFT' ) {
+      return [...acc, curr.id];
+    }
+    return [...acc];
+  }, [] );
+};
 
-  CONFIRMATION_MSG_DELAY = 3000;
+const delayUnmount = ( fn, timer, delay ) => {
+  if ( timer ) clearTimeout( timer );
+  /* eslint-disable no-param-reassign */
+  timer = setTimeout( fn, delay );
+};
 
-  componentDidMount = () => {
-    this._isMounted = true;
-  };
 
-  componentDidUpdate = () => {
-    const { actionFailures } = this.state;
-    if ( this.state.displayConfirmationMsg && !actionFailures.length ) {
-      this.delayUnmount(
-        this.hideConfirmationMsg,
-        this.confirmationMsgTimer,
-        this.CONFIRMATION_MSG_DELAY
+const TableActionsMenu = props => {
+  const {
+    loading,
+    error,
+    data: teamVideoProjects,
+    refetch: videoProjectsRefetch
+  } = useQuery( TEAM_VIDEO_PROJECTS_QUERY, {
+    variables: { ...props.variables },
+    notifyOnNetworkStatusChange: true
+  } );
+
+  const {
+    loading: projectCountLoading,
+    error: projectCountError,
+    data: teamVideoProjectsCount,
+    refetch: videoProjectsCountRefetch
+  } = useQuery(
+    TEAM_VIDEO_PROJECTS_COUNT_QUERY, {
+      variables: {
+        team: props.variables.team,
+        searchTerm: props.variables.searchTerm
+      }
+    }
+  );
+
+  const [displayConfirmationMsg, setDisplayConfirmationMsg] = useState( false );
+  const [deleteConfirmOpen, setDeleteConfirmOpen] = useState( false );
+  const [actionFailures, setActionFailures] = useState( [] );
+  const [confirmationMsgTimer, setConfirmationMsgTimer] = useState( null );
+
+  const [isMounted, setIsMounted] = useState( false );
+  useEffect( () => {
+    setIsMounted( true );
+    return () => setIsMounted( false );
+  }, [isMounted] );
+
+  useEffect( () => {
+    if ( displayConfirmationMsg && !actionFailures.length ) {
+      delayUnmount(
+        /* eslint-disable no-use-before-define */
+        hideConfirmationMsg,
+        confirmationMsgTimer,
+        CONFIRMATION_MSG_DELAY
       );
     }
+  } );
+
+  const showConfirmationMsg = () => {
+    setDisplayConfirmationMsg( true );
   };
 
-  componentWillUnmount = () => {
-    this._isMounted = false;
-    clearTimeout( this.confirmationMsgTimer );
+  const hideConfirmationMsg = () => {
+    if ( isMounted ) {
+      setDisplayConfirmationMsg( false );
+      setActionFailures( [] );
+    }
+    setConfirmationMsgTimer( null );
   };
 
-  handleActionResult = result => {
+  const handleActionResult = result => {
     if ( result.error ) {
-      this.setState( ( { actionFailures } ) => ( {
-        actionFailures: [...actionFailures, result]
-      } ) );
+      setActionFailures( prevState => [...prevState, result] );
     }
   };
 
-  handleDeleteCancel = () => {
-    this.setState( { deleteConfirmOpen: false } );
+  const handleDeleteCancel = () => {
+    setDeleteConfirmOpen( false );
   };
 
-  handleActionCompleted = () => {
-    const { variables, teamVideoProjects, teamVideoProjectsCount } = this.props;
-    teamVideoProjects.refetch( { ...variables } );
-    teamVideoProjectsCount.refetch( {
+  const handleActionCompleted = () => {
+    const { variables } = props;
+    videoProjectsRefetch( { ...variables } );
+    videoProjectsCountRefetch( {
       team: variables.team,
       searchTerm: variables.searchTerm
     } );
-    this.showConfirmationMsg();
+    showConfirmationMsg();
   };
 
-  handleStatus = ( items, projects ) => {
-    items.forEach( item => {
-      const selections = projects.filter( project => project.id === item );
-      selections.forEach( project => {
-        project.status = 'DRAFT';
-      } );
-    } );
+  const transformSelectedItemsMap = () => {
+    const { selectedItems } = props;
+    if ( selectedItems.size === 0 ) return [];
+    const arr = [];
+    selectedItems.forEach( ( value, key ) => arr.push( { id: key, value } ) );
+    return arr;
   };
 
-  getDraftProjects = projects => {
-    if ( !projects ) return [];
-    return projects.reduce( ( acc, curr ) => {
-      if ( curr.status === 'DRAFT' ) {
-        return [...acc, curr.id];
-      }
-      return [...acc];
-    }, [] );
-  };
-
-  getSelectedProjectsIds = () => {
-    const projectsArr = this.transformSelectedItemsMap();
+  const getSelectedProjectsIds = () => {
+    const projectsArr = transformSelectedItemsMap();
     if ( projectsArr && projectsArr.length ) {
       return projectsArr.reduce( ( acc, curr ) => {
         if ( curr.value ) {
@@ -106,8 +137,8 @@ class TableActionsMenu extends React.Component {
     return [];
   };
 
-  getSelectedProjects = projects => {
-    const projectIds = this.getSelectedProjectsIds();
+  const getSelectedProjects = projects => {
+    const projectIds = getSelectedProjectsIds();
     if ( projects && projects.length ) {
       const selections = projectIds.map( s => projects.find( p => s === p.id ) );
       return selections;
@@ -115,28 +146,15 @@ class TableActionsMenu extends React.Component {
     return [];
   };
 
-  getProjectsOnPage = projects => {
-    const { first, skip } = this.props.variables;
-    return projects && projects.slice( skip, skip + first );
-  };
-
-  transformSelectedItemsMap = () => {
-    const { selectedItems } = this.props;
-    if ( selectedItems.size === 0 ) return [];
-    const arr = [];
-    selectedItems.forEach( ( value, key ) => arr.push( { id: key, value } ) );
-    return arr;
-  };
-
-  hasSelectedAllDrafts = () => {
+  const hasSelectedAllDrafts = () => {
     // const draftProjects = this.getDraftProjects( this.props.teamVideoProjects.videoProjects || null );
     const teamPackages = packageMocks[0].result.data;
     const teamDocumentFiles = documentFileMocks[0].result.data;
-    const { videoProjects } = this.props.teamVideoProjects;
+    const { videoProjects } = teamVideoProjects;
     const projectsWithPackages = [...teamPackages, ...teamDocumentFiles, ...videoProjects];
-    const draftProjects = this.getDraftProjects( projectsWithPackages || null );
+    const draftProjects = getDraftProjects( projectsWithPackages || null );
 
-    const selections = this.getSelectedProjectsIds();
+    const selections = getSelectedProjectsIds();
 
     if ( selections.length > 0 ) {
       return selections.every( id => draftProjects.includes( id ) );
@@ -144,40 +162,23 @@ class TableActionsMenu extends React.Component {
     return false;
   };
 
-  showConfirmationMsg = () => {
-    this.setState( {
-      displayConfirmationMsg: true
-    } );
+  const displayConfirmDelete = () => {
+    setDeleteConfirmOpen( true );
+    setActionFailures( [] );
   };
 
-  hideConfirmationMsg = () => {
-    if ( this._isMounted ) {
-      this.setState( {
-        displayConfirmationMsg: false,
-        actionFailures: []
-      } );
-    }
-    this.confirmationMsgTimer = null;
+  const getProjectsOnPage = projects => {
+    const { first, skip } = props.variables;
+    return projects && projects.slice( skip, skip + first );
   };
 
-  delayUnmount = ( fn, timer, delay ) => {
-    if ( timer ) clearTimeout( timer );
-    /* eslint-disable no-param-reassign */
-    timer = setTimeout( fn, delay );
-  };
+  const {
+    displayActionsMenu,
+    toggleAllItemsSelection,
+    handleResetSelections
+  } = props;
 
-  displayConfirmDelete = () => {
-    this.setState( {
-      deleteConfirmOpen: true,
-      actionFailures: []
-    } );
-  };
-
-  renderMenu() {
-    const { teamVideoProjects, displayActionsMenu, toggleAllItemsSelection } = this.props;
-    const { displayConfirmationMsg, actionFailures } = this.state;
-    const { loading, error } = teamVideoProjects;
-
+  const renderMenu = () => {
     if ( loading ) return 'Loading....';
     if ( error ) return <ApolloError error={ error } />;
     if ( !teamVideoProjects || !teamVideoProjects.videoProjects ) return null;
@@ -186,19 +187,20 @@ class TableActionsMenu extends React.Component {
     // TEMP
     const teamPackages = packageMocks[0].result.data;
     const teamDocumentFiles = documentFileMocks[0].result.data;
-    const projectsWithPackages = [...teamPackages, ...teamDocumentFiles, ...videoProjects];
+    const allProjectTypes = [...teamPackages, ...teamDocumentFiles, ...videoProjects];
 
     // const projectsOnPage = this.getProjectsOnPage( videoProjects );
-    const projectsOnPage = this.getProjectsOnPage( projectsWithPackages );
+    const projectsOnPage = getProjectsOnPage( allProjectTypes );
 
     const projectsOnPageCount = getCount( projectsOnPage );
 
     // const selections = this.getSelectedProjects( videoProjects );
-    const selections = this.getSelectedProjects( projectsWithPackages );
+    const selections = getSelectedProjects( allProjectTypes );
 
     const selectionsCount = getCount( selections );
 
-    const isDisabled = videoProjects && !videoProjects.length;
+    // const isDisabled = videoProjects && !videoProjects.length;
+    const isDisabled = allProjectTypes && !allProjectTypes.length;
 
     const isChecked = projectsOnPageCount === selectionsCount && selectionsCount > 0;
 
@@ -222,7 +224,7 @@ class TableActionsMenu extends React.Component {
           <Modal
             className="confirmation"
             closeIcon
-            onClose={ this.hideConfirmationMsg }
+            onClose={ hideConfirmationMsg }
             open={ displayConfirmationMsg }
             size="tiny"
           >
@@ -237,16 +239,16 @@ class TableActionsMenu extends React.Component {
             <img src={ editIcon } alt="Edit Selection(s)" title="Edit Selection(s)" />
           </Button>
 
-          <DeleteIconButton displayConfirmDelete={ this.displayConfirmDelete } />
+          <DeleteIconButton displayConfirmDelete={ displayConfirmDelete } />
 
           <DeleteProjects
-            deleteConfirmOpen={ this.state.deleteConfirmOpen }
-            handleDeleteCancel={ this.handleDeleteCancel }
-            handleDeleteConfirm={ this.handleActionCompleted }
-            handleActionResult={ this.handleActionResult }
-            handleResetSelections={ this.props.handleResetSelections }
+            deleteConfirmOpen={ deleteConfirmOpen }
+            handleDeleteCancel={ handleDeleteCancel }
+            handleDeleteConfirm={ handleActionCompleted }
+            handleActionResult={ handleActionResult }
+            handleResetSelections={ handleResetSelections }
             selections={ selections }
-            showConfirmationMsg={ this.showConfirmationMsg }
+            showConfirmationMsg={ showConfirmationMsg }
           />
 
           <Button size="mini" basic disabled>
@@ -256,13 +258,13 @@ class TableActionsMenu extends React.Component {
             <img src={ archiveIcon } alt="Archive Selection(s)" title="Archive Selection(s)" />
           </Button>
 
-          { !this.hasSelectedAllDrafts() && (
+          { !hasSelectedAllDrafts() && (
             <>
               <span className="separator">|</span>
               <UnpublishProjects
-                handleResetSelections={ this.props.handleResetSelections }
-                handleActionResult={ this.handleActionResult }
-                showConfirmationMsg={ this.showConfirmationMsg }
+                handleResetSelections={ handleResetSelections }
+                handleActionResult={ handleActionResult }
+                showConfirmationMsg={ showConfirmationMsg }
                 selections={ selections }
               />
             </>
@@ -270,38 +272,17 @@ class TableActionsMenu extends React.Component {
         </div>
       </Fragment>
     );
-  }
+  };
 
-  render() {
-    return <div className="actionsMenu_wrapper">{ this.renderMenu() }</div>;
-  }
-}
+  return <div className="actionsMenu_wrapper">{ renderMenu() }</div>;
+};
 
 TableActionsMenu.propTypes = {
   displayActionsMenu: PropTypes.bool,
   variables: PropTypes.object,
   selectedItems: PropTypes.object,
-  teamVideoProjects: PropTypes.object,
-  teamVideoProjectsCount: PropTypes.object,
   handleResetSelections: PropTypes.func,
   toggleAllItemsSelection: PropTypes.func
 };
 
-export default compose(
-  graphql( TEAM_VIDEO_PROJECTS_QUERY, {
-    name: 'teamVideoProjects',
-    options: props => ( {
-      variables: { ...props.variables },
-      notifyOnNetworkStatusChange: true
-    } )
-  } ),
-  graphql( TEAM_VIDEO_PROJECTS_COUNT_QUERY, {
-    name: 'teamVideoProjectsCount',
-    options: props => ( {
-      variables: {
-        team: props.variables.team,
-        searchTerm: props.variables.searchTerm
-      }
-    } )
-  } )
-)( TableActionsMenu );
+export default TableActionsMenu;
