+++ components/admin/Dashboard/TeamProjects/DetailsPopup/DetailsPopup.js
@@ -1,12 +1,15 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import dynamic from 'next/dynamic';
-import { Query } from 'react-apollo';
+import { useQuery } from '@apollo/react-hooks';
 import gql from 'graphql-tag';
 import debounce from 'lodash/debounce';
 import { Popup } from 'semantic-ui-react';
 import ApolloError from 'components/errors/ApolloError';
 import './DetailsPopup.scss';
+
+const VideoDetailsPopup = dynamic( () => import( './VideoDetailsPopup' ) );
+const PackageDetailsPopup = dynamic( () => import( './PackageDetailsPopup' ) );
 
 const CHECK_PROJECT_TYPE_QUERY = gql`
   query CheckProjectType( $id: ID! ) {
@@ -17,15 +20,24 @@ const CHECK_PROJECT_TYPE_QUERY = gql`
   }
 `;
 
-const VideoDetailsPopup = dynamic( () => import( './VideoDetailsPopup' ) );
-const ImageDetailsPopup = dynamic( () => import( './ImageDetailsPopup' ) );
+// const CHECK_TYPE_QUERY = gql`
+//   query CheckTypeQuery( $id: ID! ) {
+//     videoProject( id: $id ) {
+//       id
+//       projectType
+//     }
+//     package( id: $id ) {
+//       id
+//     }
+//   }
+// `;
 
 const renderPopup = ( projectType, id ) => {
   if ( projectType === 'VideoProject' ) {
     return <VideoDetailsPopup id={ id } />;
   }
-  if ( projectType === 'ImageProject' ) {
-    return <ImageDetailsPopup id={ id } />;
+  if ( projectType === 'Package' ) {
+    return <PackageDetailsPopup id={ id } />;
   }
 };
 
@@ -68,42 +80,50 @@ const DetailsPopup = props => {
     }
   };
 
-  return (
-    <Query query={ CHECK_PROJECT_TYPE_QUERY } variables={ { id: props.id } }>
-      {
-        ( { loading, error, data } ) => {
-          if ( loading ) return <p>Loading....</p>;
-          if ( error ) return <ApolloError error={ error } />;
-          if ( !data.videoProject ) return null;
+  const { loading, error, data } = useQuery( CHECK_PROJECT_TYPE_QUERY, {
+    variables: { id: props.id }
+  } );
 
-          const { __typename } = data.videoProject;
-          return (
-            // 06/10/19 - Updating button text from "Details" to "Files"
-            // if DetailsPopup will contain content other than files in future,
-            // will add conditional to display "Details" text along with "Files"/"Other Content"
-            // subheaders within popup
-            <Popup
-              className="detailsFiles_popup"
-              trigger={ (
-                <button
-                  type="button"
-                  className="linkStyle projects_data_actions_action"
-                  data-projectitempopup="detailsPopup"
-                >
-                  Files
-                </button>
-              ) }
-              content={ renderPopup( __typename, id ) }
-              on="click"
-              position="bottom left"
-              open={ detailsPopupOpen }
-              onOpen={ handleOpen }
-              onClose={ handleClose }
-            />
-          );
-        }
-      }
-    </Query>
+  if ( loading ) return <p>Loading....</p>;
+  if ( error ) return <ApolloError error={ error } />;
+  // if ( !data.videoProject && !data.package ) return null;
+
+  // TEMP
+  if ( id === 'package123' ) {
+    data.package = {
+      __typename: 'Package',
+      id,
+    };
+  }
+
+  // const { __typename } = data.videoProject;
+  let __typename;
+  if ( data.videoProject != null ) __typename = data.videoProject.__typename;
+  if ( data.package != null ) __typename = data.package.__typename;
+
+  return (
+    // 06/10/19 - Updating button text from "Details" to "Files"
+    // if DetailsPopup will contain content other than files in future,
+    // will add conditional to display "Details" text along with "Files"/"Other Content"
+    // subheaders within popup
+    <Popup
+      className="detailsFiles_popup"
+      trigger={ (
+        <button
+          type="button"
+          className="linkStyle projects_data_actions_action"
+          data-projectitempopup="detailsPopup"
+        >
+          Files
+        </button>
+      ) }
+      content={ renderPopup( __typename, id ) }
+      on="click"
+      position="bottom left"
+      open={ detailsPopupOpen }
+      onOpen={ handleOpen }
+      onClose={ handleClose }
+    />
   );
 };
 
@@ -112,5 +132,3 @@ DetailsPopup.propTypes = {
 };
 
 export default DetailsPopup;
-
-export { CHECK_PROJECT_TYPE_QUERY };
